[mem_cache][6/N] refactor: move MHA host-pool family into pool_host/mha.py
