+++ test/registered/jit/test_kvcacheio_asymmetric.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 import pytest
 import torch
 
-from sglang.srt.mem_cache.memory_pool_host import AsymmetricMHATokenToKVPoolHost
+from sglang.srt.mem_cache.pool_host.mha import AsymmetricMHATokenToKVPoolHost
 from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 
 register_cuda_ci(est_time=10, stage="base-b-kernel-unit", runner_config="1-gpu-large")
