+++ test/registered/unit/mem_cache/test_hicache_staged_write_back_dispatch.py
@@ -25,15 +25,16 @@ from sglang.srt.mem_cache.memory_pool_host import (
     HostPoolGroup,
     LogicalHostPool,
     MambaPoolHost,
-    MHATokenToKVPoolHost,
     MLATokenToKVPoolHost,
     PoolEntry,
 )
+from sglang.srt.mem_cache.pool_host.mha import MHATokenToKVPoolHost
 from sglang.test.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=3, suite="base-a-test-cpu")
 
 MEMORY_POOL_HOST_MODULE = "sglang.srt.mem_cache.memory_pool_host"
+MHA_POOL_HOST_MODULE = "sglang.srt.mem_cache.pool_host.mha"
 
 
 def _indices(start: int, end: int) -> torch.Tensor:
@@ -230,21 +231,21 @@ class TestHiCacheStagedWriteBackDispatch(unittest.TestCase):
 
         with (
             mock.patch(
-                f"{MEMORY_POOL_HOST_MODULE}.jit_transfer_hicache_all_layer_staged_lf_pf",
+                f"{MHA_POOL_HOST_MODULE}.jit_transfer_hicache_all_layer_staged_lf_pf",
                 side_effect=lambda **kwargs: _cpu_staged_mha_lf_pf_copy(
                     src_registry, **kwargs
                 ),
             ) as staged,
             mock.patch(
-                f"{MEMORY_POOL_HOST_MODULE}.transfer_kv_all_layer_lf_pf",
+                f"{MHA_POOL_HOST_MODULE}.transfer_kv_all_layer_lf_pf",
                 create=True,
             ) as fallback,
             mock.patch(
-                f"{MEMORY_POOL_HOST_MODULE}.jit_transfer_hicache_one_layer",
+                f"{MHA_POOL_HOST_MODULE}.jit_transfer_hicache_one_layer",
                 side_effect=_cpu_jit_one_layer_mha_copy,
             ) as load,
             mock.patch(
-                f"{MEMORY_POOL_HOST_MODULE}.can_use_write_back_jit_kernel",
+                f"{MHA_POOL_HOST_MODULE}.can_use_write_back_jit_kernel",
                 return_value=True,
             ) as can_use_write_back_jit_kernel,
         ):
